backend/requirements.txt
# Core dependencies
Flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
Flask-Limiter==3.5.0

# Utilities
requests==2.31.0
openpyxl==3.1.2
python-dotenv==1.0.0
cryptography==42.0.8
reportlab==4.2.5
coverage==7.13.4
mysql-connector-python==9.0.0
pyotp==2.9.0

# Security
bcrypt==4.1.2

# Production server
gunicorn==21.2.0

# Database
# SQLite is included in Python standard library
